Implement Iterator for RopeSlice
This patch adds a new macro called generate_ropeslice_struct which is
used in both the rope & src_rope modules. It takes as an argument the
Lnode type defined in each of the aforementioned modules. The two
definitions of Lnode are slightly different, which require two different
definitions of RopeSlice, which is why I've done things like this.